Cleaned up overlays
