+++ lib/components/Alert/index.js
@@ -5,22 +5,23 @@ import { AnimatePresence, motion } from "framer-motion";
 import { useHotkeys } from "react-hotkeys-hook";
 import { Close } from "@bigbinary/neeto-icons";
 
-import { Portal, Backdrop } from "../atoms";
-import { useOnClickOutside } from "../utils";
-import { BLUR_INITIAL, BLUR_FINAL, TRANSITION } from "../constants/overlay";
-import Button from "./Button";
-import Typography from "./Typography";
-import PageLoader from "./PageLoader";
+import { Portal, Backdrop } from "../../atoms";
+import { useOnClickOutside } from "../../utils";
+import { BLUR_INITIAL, BLUR_FINAL, TRANSITION } from "../../constants/overlay";
+import Button from "../Button";
+import PageLoader from "../PageLoader";
+import Body from "./Body";
 
 const noop = () => {};
-const sizes = {
+
+const SIZES = {
   xs: "xs",
   sm: "sm",
   md: "md",
 };
 
 const Alert = ({
-  size = "sm",
+  size = SIZES.sm,
   isOpen = false,
   loading = false,
   isSubmitting = false,
@@ -71,9 +72,9 @@ const Alert = ({
                 key="modal-wrapper"
                 exit={{ scale: 1, opacity: 0, y: 20, filter: BLUR_FINAL }}
                 className={classnames("neeto-ui-modal__wrapper", {
-                  "neeto-ui-modal__wrapper--xs": size === sizes.xs,
-                  "neeto-ui-modal__wrapper--sm": size === sizes.sm,
-                  "neeto-ui-modal__wrapper--md": size === sizes.md,
+                  "neeto-ui-modal__wrapper--xs": size === SIZES.xs,
+                  "neeto-ui-modal__wrapper--sm": size === SIZES.sm,
+                  "neeto-ui-modal__wrapper--md": size === SIZES.md,
                   [className]: className,
                 })}
               >
@@ -112,7 +113,7 @@ Alert.propTypes = {
   /**
    * To specify the size of the Alert.
    */
-  size: PropTypes.oneOf(Object.values(sizes)),
+  size: PropTypes.oneOf(Object.values(SIZES)),
   /**
    * To specify whether the Alert is open or not.
    */
@@ -154,46 +155,6 @@ Alert.propTypes = {
   cancelButtonLabel: PropTypes.string,
   backdropClassName: PropTypes.string,
   closeOnOutsideClick: PropTypes.bool,
-};
-
-const Body = ({
-  title,
-  message,
-  submitButtonLabel,
-  cancelButtonLabel,
-  onSubmit,
-  onClose,
-  isSubmitting,
-}) => {
-  return (
-    <>
-      <div className="neeto-ui-modal__header" data-cy="alert-title">
-        {title && <Typography style="h2">{title}</Typography>}
-      </div>
-      <div className="neeto-ui-modal__body" data-cy="alert-message">
-        <Typography style="body2" lineHeight="normal">
-          {message}
-        </Typography>
-      </div>
-      <div className="space-x-2 neeto-ui-alert__footer">
-        <Button
-          style="danger"
-          size="large"
-          data-cy="alert-submit-button"
-          label={submitButtonLabel}
-          onClick={onSubmit}
-          loading={isSubmitting}
-        />
-        <Button
-          style="text"
-          size="large"
-          data-cy="alert-cancel-button"
-          label={cancelButtonLabel}
-          onClick={onClose}
-        />
-      </div>
-    </>
-  );
 };
 
 export default Alert;
