+++ lib/components/Modal/index.js
@@ -5,26 +5,30 @@ import { AnimatePresence, motion } from "framer-motion";
 import { useHotkeys } from "react-hotkeys-hook";
 import { Close } from "@bigbinary/neeto-icons";
 
-import Button from "./Button";
-import { Portal, Backdrop } from "../atoms";
-import { useOnClickOutside } from "../utils";
-import { BLUR_INITIAL, BLUR_FINAL, TRANSITION } from "../constants/overlay";
+import { Portal, Backdrop } from "../../atoms";
+import { useOnClickOutside } from "../../utils";
+import { BLUR_INITIAL, BLUR_FINAL, TRANSITION } from "../../constants/overlay";
+import Button from "../Button";
+import Header from "./Header";
+import Body from "./Body";
+import Footer from "./Footer";
 
 const noop = () => {};
-const sizes = {
+
+const SIZES = {
   xs: "xs",
   sm: "sm",
   md: "md",
 };
 
-const animationVariants = {
+const ANIMATION_VARIANTS = {
   open: { scale: 1, opacity: 1, y: 0, filter: BLUR_INITIAL },
   closed: { scale: 1.1, opacity: 0, y: -20, filter: BLUR_FINAL },
   exit: { scale: 1, opacity: 0, y: 20, filter: BLUR_FINAL },
 };
 
 const Modal = ({
-  size = "sm",
+  size = SIZES.sm,
   isOpen = false,
   onClose = noop,
   loading = false,
@@ -62,7 +66,7 @@ const Modal = ({
             <motion.div
               role="dialog"
               aria-modal={true}
-              variants={animationVariants}
+              variants={ANIMATION_VARIANTS}
               initial="closed"
               animate="open"
               exit="exit"
@@ -73,9 +77,9 @@ const Modal = ({
               ref={modalWrapper}
               key="modal-wrapper"
               className={classnames("neeto-ui-modal__wrapper", {
-                "neeto-ui-modal__wrapper--xs": size === sizes.xs,
-                "neeto-ui-modal__wrapper--sm": size === sizes.sm,
-                "neeto-ui-modal__wrapper--md": size === sizes.md,
+                "neeto-ui-modal__wrapper--xs": size === SIZES.xs,
+                "neeto-ui-modal__wrapper--sm": size === SIZES.sm,
+                "neeto-ui-modal__wrapper--md": size === SIZES.md,
                 [className]: className,
               })}
               {...otherProps}
@@ -102,7 +106,7 @@ Modal.propTypes = {
   /**
    * To specify the size of the Modal.
    */
-  size: PropTypes.oneOf(Object.values(sizes)),
+  size: PropTypes.oneOf(Object.values(SIZES)),
   /**
    * To specify whether the Modal is open or not.
    */
@@ -124,30 +128,6 @@ Modal.propTypes = {
   closeButton: PropTypes.bool,
   backdropClassName: PropTypes.string,
   closeOnOutsideClick: PropTypes.bool,
-};
-
-const Header = ({ children, className }) => {
-  return (
-    <div className={classnames("neeto-ui-modal__header", className)}>
-      {children}
-    </div>
-  );
-};
-
-const Body = ({ children, className }) => {
-  return (
-    <div className={classnames("neeto-ui-modal__body", className)}>
-      {children}
-    </div>
-  );
-};
-
-const Footer = ({ children, className }) => {
-  return (
-    <div className={classnames("neeto-ui-modal__footer", className)}>
-      {children}
-    </div>
-  );
 };
 
 Modal.Header = Header;
