+++ lib/components/Pane/index.js
@@ -5,20 +5,24 @@ import { AnimatePresence, motion } from "framer-motion";
 import { useHotkeys } from "react-hotkeys-hook";
 import classnames from "classnames";
 
-import Button from "./Button";
-import { Portal, Backdrop } from "../atoms";
-import { useOnClickOutside } from "../utils";
+import { Portal, Backdrop } from "../../atoms";
+import { useOnClickOutside } from "../../utils";
 import {
   PANE_SIZE,
   BLUR_INITIAL,
   BLUR_FINAL,
   TRANSITION,
-} from "../constants/overlay";
+} from "../../constants/overlay";
+import Button from "../Button";
+import Header from "./Header";
+import Body from "./Body";
+import Footer from "./Footer";
 
 const noop = () => {};
-const sizes = ["sm", "lg"];
 
-const animationVariants = {
+const SIZES = ["sm", "lg"];
+
+const ANIMATION_VARIANTS = {
   open: {
     x: 0,
     filter: BLUR_INITIAL,
@@ -31,7 +35,7 @@ const animationVariants = {
 
 const Pane = ({
   size = "sm",
-  isOpen,
+  isOpen = false,
   onClose = noop,
   loading = false,
   children,
@@ -44,6 +48,7 @@ const Pane = ({
 }) => {
   const paneWrapper = useRef();
   const backdropRef = useRef();
+  
   useOnClickOutside(
     paneWrapper,
     backdropRef,
@@ -65,7 +70,7 @@ const Pane = ({
             )}
           >
             <motion.div
-              variants={animationVariants}
+              variants={ANIMATION_VARIANTS}
               initial="closed"
               animate="open"
               exit="closed"
@@ -102,7 +107,7 @@ const Pane = ({
 };
 
 Pane.propTypes = {
-  size: PropTypes.oneOf(sizes),
+  size: PropTypes.oneOf(SIZES),
   /**
    * To specify whether the pane component is open or not.
    */
@@ -136,72 +141,6 @@ Pane.propTypes = {
    * To specify whether the pane component should close on outside click.
    */
   closeOnOutsideClick: PropTypes.bool,
-};
-
-export const Header = ({ children, className }) => {
-  return (
-    <div className={classnames("neeto-ui-pane__header", className)}>
-      {children}
-    </div>
-  );
-};
-
-Header.propTypes = {
-  /**
-   * To specify className to be applied to the pane header container.
-   */
-  className: PropTypes.string,
-  children: PropTypes.node,
-};
-
-export const Body = ({ children, className, hasFooter = true }) => {
-  return (
-    <div
-      className={classnames(
-        "neeto-ui-pane__body flex flex-col items-start justify-start",
-        {
-          "neeto-ui-pane__body--has-footer": hasFooter,
-          [className]: className,
-        }
-      )}
-    >
-      {children}
-    </div>
-  );
-};
-
-Body.propTypes = {
-  /**
-   * To specify if the pane has a footer.
-   * @default true
-   */
-  hasFooter: PropTypes.bool,
-  /**
-   * To specify className to be applied to the pane body container.
-   */
-  className: PropTypes.string,
-  children: PropTypes.node,
-};
-
-export const Footer = ({ children, className }) => {
-  return (
-    <div
-      className={classnames(
-        "neeto-ui-pane__footer neeto-ui-shadow-m flex items-center",
-        className
-      )}
-    >
-      {children}
-    </div>
-  );
-};
-
-Footer.propTypes = {
-  /**
-   * To specify className to be applied to the pane footer container.
-   */
-  className: PropTypes.string,
-  children: PropTypes.node,
 };
 
 Pane.Header = Header;
