+++ lib/components/Toastr/index.js
@@ -1,6 +1,5 @@
 import React from "react";
 import { join } from "ramda";
-import Linkify from "react-linkify";
 import { toast, Slide } from "react-toastify";
 import {
   CheckCircle,
@@ -10,9 +9,8 @@ import {
   Close,
 } from "@bigbinary/neeto-icons";
 
-import "../styles/components/_toast.scss";
-
-const noop = () => {};
+import "../../styles/components/_toast.scss";
+import Toast from "./Toast";
 
 const TOAST_CONFIG = {
   autoClose: 3500,
@@ -37,44 +35,9 @@ const TOAST_ICON = {
   info: <Info color="#276EF1" />,
 };
 
-const ToastrComponent = ({
-  type,
-  message,
-  buttonLabel = "",
-  onClick = noop,
-}) => {
-  return (
-    <div
-      data-cy="toastr-message-container"
-      data-test={`toastr-${type}-container`}
-    >
-      <Linkify
-        componentDecorator={(decoratedHref, decoratedText, key) => (
-          <a target="_blank" href={decoratedHref} key={key} rel="noreferrer">
-            {decoratedText}
-          </a>
-        )}
-      >
-        <p>{message}</p>
-      </Linkify>
-      {buttonLabel && (
-        <button
-          onClick={(e) => {
-            e.stopPropagation();
-            onClick();
-          }}
-          data-cy={`toastr-${type}-button`}
-        >
-          {buttonLabel}
-        </button>
-      )}
-    </div>
-  );
-};
-
 const showSuccessToastr = (message, buttonLabel, onClick) => {
   toast.success(
-    <ToastrComponent
+    <Toast
       type="success"
       message={message}
       buttonLabel={buttonLabel}
@@ -89,7 +52,7 @@ const showSuccessToastr = (message, buttonLabel, onClick) => {
 
 const showInfoToastr = (message, buttonLabel, onClick) => {
   toast.info(
-    <ToastrComponent
+    <Toast
       type="info"
       message={message}
       buttonLabel={buttonLabel}
@@ -101,7 +64,7 @@ const showInfoToastr = (message, buttonLabel, onClick) => {
 
 const showWarningToastr = (message, buttonLabel, onClick) => {
   toast.warning(
-    <ToastrComponent
+    <Toast
       type="warning"
       message={message}
       buttonLabel={buttonLabel}
@@ -120,15 +83,17 @@ const isString = (s) => typeof s === "string" || s instanceof String;
 
 const showErrorToastr = (errorObject, buttonLabel, onClick) => {
   let errorMessage;
+
   if (isAxiosError(errorObject)) {
     const { error, errors } = errorObject.response.data;
-    errorMessage = error || (errors && join("\n", errors)) || errorObject.message;
-  }
-  else if (isError(errorObject)) errorMessage = errorObject.message;
+    errorMessage =
+      error || (errors && join("\n", errors)) || errorObject.message;
+  } else if (isError(errorObject)) errorMessage = errorObject.message;
   else if (isString(errorObject)) errorMessage = errorObject;
   else errorMessage = "Something went wrong.";
+
   toast.error(
-    <ToastrComponent
+    <Toast
       type="error"
       message={errorMessage}
       buttonLabel={buttonLabel}
@@ -143,7 +108,7 @@ const showErrorToastr = (errorObject, buttonLabel, onClick) => {
   );
 };
 
-export const Toastr = {
+const Toastr = {
   info: showInfoToastr,
   show: showInfoToastr,
   success: showSuccessToastr,
