+++ lib/components/Tooltip.js
@@ -3,6 +3,8 @@ import PropTypes from "prop-types";
 import Tippy from "@tippyjs/react";
 import { followCursor } from "tippy.js";
 
+import { ARROW } from "../constants/overlay";
+
 const Tooltip = ({
   content,
   children,
@@ -15,18 +17,18 @@ const Tooltip = ({
   ...otherProps
 }) => {
   const localProps = {};
+
   if (hideAfter > 0) {
-    localProps["onShow"] = instance => setTimeout(() => instance.hide(), hideAfter);
-  };
+    localProps["onShow"] = (instance) =>
+      setTimeout(() => instance.hide(), hideAfter);
+  }
 
   return (
     <Tippy
       role="tooltip"
       theme={theme}
       content={content}
-      arrow={
-        "<svg width='12' height='6' viewBox='0 0 10 5' fill='none' xmlns='http://www.w3.org/2000/svg'><path d='M10 5H0.926697L3.95208 1.63847C4.74227 0.760478 6.11722 0.754951 6.91445 1.62656L10 5Z' /></svg>"
-      }
+      arrow={ARROW}
       disabled={disabled}
       animation={"scale-subtle"}
       placement={placement || position}
